chore: various cleaning and quick fixes
