+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2025-02-11
+## [1.0.0] - 2026-04-XX
+
+### BREAKING
+Complete rewrite of the starter, upgraded dependencies, new features,… too many changes to list them all. Check the [wiki](https://github.com/GoOz/Niepce/wiki) for more informations and [MIGRATION.md](MIGRATION.md) for migration intructions.
+
+## [0.5.0] - 2026-03-29
+
+### Add
+
+- Add Footer informations
+- Add trigger warning flag
+
+### Fix
+- Update Glass logo
+
+## [0.4.0] - 2025-02-11
 
 ### Add
