+++ _includes/layouts/archives.njk
@@ -1,14 +1,12 @@
 ---
 layout: layouts/base.njk
 ---
-<articile class="article">
-  <h1 class="article-title">{{ title }}</h1>
+<h1 class="article-title">{{ title }}</h1>
 
-  <div class="article-content centered">
-    {{ content | safe }}
-  </div>
+<div class="article-content centered">
+  {{ content | safe }}
+</div>
 
-  {% set posts = collections.post %}
-  {% include "get_posts.njk" %}
-</articile>
+{% set posts = collections.post %}
+{% include "get_posts.njk" %}
