+++ _includes/layouts/blogpost.njk
@@ -1,7 +1,7 @@
 ---
 layout: layouts/base.njk
 ---
-<article class="article">
+<article>
 	<h1 class="article-title">{{ title }}</h1>
 
 	<div class="article-content">
