+++ _includes/layouts/category_list.njk
@@ -1,13 +1,11 @@
 ---
 layout: layouts/base.njk
 ---
-<article class="article">
-  <h1 class="article-title">{{ title }}</h1>
+<h1 class="article-title">{{ title }}</h1>
 
-  <div class="article-content centered">
-    {{ content | safe }}
-  </div>
+<div class="article-content centered">
+  {{ content | safe }}
+</div>
 
-  {% set posts = collections.post %}
-  {% include "get_categories.njk" %}
-</article>
+{% set posts = collections.post %}
+{% include "get_categories.njk" %}
