+++ _includes/layouts/page.njk
@@ -2,9 +2,7 @@
 layout: layouts/base.njk
 ---
 
-<article class="article">
-	<h1 class="article-title">{{ title }}</h1>
+<h1 class="article-title">{{ title }}</h1>
 
-	<div class="article-content">{{ content | safe }}</div>
+<div class="article-content">{{ content | safe }}</div>
 
-</article>
