+++ _includes/layouts/post.njk
@@ -1,7 +1,7 @@
 ---
 layout: layouts/base.njk
 ---
-<article class="article">
+<article>
 	{% if photo %}
 	<figure class="photo">
 		<img src={{ photo }} alt="{{photo_alt}}"
