+++ _includes/layouts/series.njk
@@ -1,13 +1,11 @@
 ---
 layout: layouts/base.njk
 ---
-<articile class="article">
-  <h1 class="article-title">{{ title }}</h1>
+<h1 class="article-title">{{ title }}</h1>
 
-  <div class="article-content centered">
-    {{ content | safe }}
-  </div>
+<div class="article-content centered">
+  {{ content | safe }}
+</div>
 
-  {% set posts = collections.series | filterBySeries(seriesName) %}
-  {% include "get_posts.njk" %}
-</articile>
+{% set posts = collections.series | filterBySeries(seriesName) %}
+{% include "get_posts.njk" %}
