+++ _includes/layouts/series_list.njk
@@ -1,13 +1,11 @@
 ---
 layout: layouts/base.njk
 ---
-<articile class="article">
-  <h1 class="article-title">{{ title }}</h1>
+<h1 class="article-title">{{ title }}</h1>
 
-  <div class="article-content centered">
-    {{ content | safe }}
-  </div>
+<div class="article-content centered">
+  {{ content | safe }}
+</div>
 
-  {% set posts = collections.series %}
-  {% include "get_series.njk" %}
-</articile>
+{% set posts = collections.series %}
+{% include "get_series.njk" %}
