+++ content/pages/archives.md
@@ -7,5 +7,3 @@ override:tags: ["pages"]
 title: All my pictures
 permalink: /archives/
 ---
-
-**All my pictures so far…**
