+++ public/css/niepce.css
@@ -680,7 +680,7 @@ a.logo {
 }
 
 /* Network block */
-.network {
+:is(.network, .blogposts-title) {
 	box-sizing: border-box;
 	display: flex;
 	width: 100%;
@@ -759,7 +759,7 @@ a.logo {
     border-radius: 0;
   }
 }
-:where(.grid:not(.monocolumn), .article) {
+:where(.grid:not(.monocolumn)) {
   & .post-title {
     position: absolute;
     inset: 0;
@@ -1100,11 +1100,11 @@ a.logo {
 }
 .footer-nav-prev {
 	clip-path: polygon(20% 0, 100% 0%, 100% 100%, 20% 100%, 0 50%);
-	margin-block-end: auto;
+	margin-inline-end: auto;
 }
 .footer-nav-next {
 	clip-path: polygon(0 0, 80% 0%, 100% 50%, 80% 100%, 0 100%);
-	margin-block-start: auto;
+	margin-inline-start: auto;
 }
 @media (orientation: portrait), (max-width: 40em) {
 	.footer-nav > * {
@@ -1151,8 +1151,7 @@ a.logo {
     align-items: center;
     justify-content: center;
     margin: 0;
-    background-color: var(--overlay);
-    color: var(--background);
+    background-color: var(--splash-overlay);
   }
 }
 
@@ -1269,6 +1268,13 @@ button {
   }
 }
 
+.blogposts-title {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--size-3);
+}
+
 .blog-nav {
   display: flex;
   flex-wrap: nowrap;
@@ -1303,8 +1309,14 @@ button {
     flex-shrink: 0;
   }
 
-  & .blog-nav-prev svg {
-    transform: rotate(180deg);
+  & .blog-nav-next {
+    margin-inline-start: auto;
+  }
+  & .blog-nav-prev {
+    margin-inline-end: auto;
+    & svg {
+      transform: rotate(180deg);
+    }
   }
 }
 @media (orientation: portrait), (max-width: 40em) {
